Feature: countdown before switching back to broadcasting mode
This feature was implemented as a one new publication channel
(`countdown_to_end_split_mode`).  It's issued when the instructor hits
"Back to broadcasting mode" button.

Then, in instructor's browser, the countdown starts.  After 15-30 seconds
(subject to change) the "end_split_mode" RPC is called.  Previously it'd
happen when the switch mode button was pressed.